+++ tests/test_ade20k2017_format.py
@@ -9,6 +9,7 @@ import numpy as np
 
 from datumaro.components.annotation import AnnotationType, LabelCategories, Mask
 from datumaro.components.dataset import Dataset, DatasetItem
+from datumaro.components.environment import Environment
 from datumaro.plugins.ade20k2017_format import Ade20k2017Importer
 from datumaro.util.test_utils import compare_datasets
 
@@ -19,7 +20,8 @@ DUMMY_DATASET_DIR = osp.join(osp.dirname(__file__), 'assets', 'ade20k2017_datase
 class Ade20k2017ImporterTest(TestCase):
     @mark_requirement(Requirements.DATUM_399)
     def test_can_detect(self):
-        self.assertTrue(Ade20k2017Importer.detect(DUMMY_DATASET_DIR))
+        detected_formats = Environment().detect_dataset(DUMMY_DATASET_DIR)
+        self.assertIn(Ade20k2017Importer.NAME, detected_formats)
 
     @mark_requirement(Requirements.DATUM_399)
     def test_can_import(self):