+++ tests/test_kitti_format.py
@@ -9,6 +9,7 @@ from datumaro.components.annotation import (
     AnnotationType, Bbox, LabelCategories, Mask,
 )
 from datumaro.components.dataset import Dataset
+from datumaro.components.environment import Environment
 from datumaro.components.extractor import DatasetItem, Extractor
 from datumaro.plugins.kitti_format.converter import KittiConverter
 from datumaro.plugins.kitti_format.format import (
@@ -127,9 +128,12 @@ class KittiImportTest(TestCase):
                 KittiSegmentationImporter),
         ]
 
+        env = Environment()
+
         for path, subtask in matrix:
             with self.subTest(path=path, task=subtask):
-                self.assertTrue(subtask.detect(path))
+                detected_formats = env.detect_dataset(path)
+                self.assertIn(subtask.NAME, detected_formats)
 
 
 class TestExtractorBase(Extractor):