+++ tests/test_labelme_format.py
@@ -7,6 +7,7 @@ import numpy as np
 
 from datumaro.components.annotation import Bbox, Mask, Polygon
 from datumaro.components.dataset import Dataset
+from datumaro.components.environment import Environment
 from datumaro.components.extractor import DatasetItem
 from datumaro.plugins.labelme_format import LabelMeConverter, LabelMeImporter
 from datumaro.util.image import Image
@@ -184,7 +185,8 @@ DUMMY_DATASET_DIR = osp.join(osp.dirname(__file__), 'assets', 'labelme_dataset')
 class LabelMeImporterTest(TestCase):
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_can_detect(self):
-        self.assertTrue(LabelMeImporter.detect(DUMMY_DATASET_DIR))
+        detected_formats = Environment().detect_dataset(DUMMY_DATASET_DIR)
+        self.assertIn(LabelMeImporter.NAME, detected_formats)
 
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_can_import(self):