+++ tests/test_mnist_csv_format.py
@@ -7,6 +7,7 @@ from datumaro.components.annotation import (
     AnnotationType, Label, LabelCategories,
 )
 from datumaro.components.dataset import Dataset
+from datumaro.components.environment import Environment
 from datumaro.components.extractor import DatasetItem
 from datumaro.plugins.mnist_csv_format import (
     MnistCsvConverter, MnistCsvImporter,
@@ -197,4 +198,5 @@ class MnistCsvImporterTest(TestCase):
 
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_can_detect(self):
-        self.assertTrue(MnistCsvImporter.detect(DUMMY_DATASET_DIR))
+        detected_formats = Environment().detect_dataset(DUMMY_DATASET_DIR)
+        self.assertIn(MnistCsvImporter.NAME, detected_formats)